ref: use dedicated modal component
