+++ AMW_angular/io/src/app/resources/resource-edit/resource-edit.component.html
@@ -91,5 +91,4 @@
       </div>
     }
   </div>
-  <app-tag-current-state [resource]="resource()"></app-tag-current-state>
 </app-page>
